# WebApp/swagger.json
{
  "x-generator": "NSwag v13.11.3.0 (NJsonSchema v10.4.4.0 (Newtonsoft.Json v12.0.0.0))",
  "openapi": "3.0.0",
  "info": {
    "title": "checkbiz-client-service",
    "version": "1.0.0"
  },
  "paths": {
    "/norway-lookup/ssn/{ssn}": {
      "get": {
        "tags": [
          "NorwayLookup"
        ],
        "operationId": "LookupBySsn",
        "parameters": [
          {
            "name": "ssn",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LookupIdNoResult"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "LookupIdNoResult": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "responseCode": {
            "type": "string"
          },
          "count": {
            "type": "integer",
            "format": "int32"
          },
          "contacts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Contact"
            }
          }
        }
      },
      "Contact": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "geography": {
            "$ref": "#/components/schemas/Geography"
          },
          "contactPoints": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ContactPoint"
            }
          }
        }
      },
      "Geography": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "street": {
            "type": "string"
          },
          "houseNumber": {
            "type": "string"
          },
          "entrance": {
            "type": "string"
          },
          "postCode": {
            "type": "string"
          },
          "postArea": {
            "type": "string"
          },
          "addressString": {
            "type": "string"
          }
        }
      },
      "ContactPoint": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "label": {
            "type": "string"
          },
          "main": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "value": {
            "type": "string"
          }
        }
      }
    }
  }
}